Created eventHandler.h
eventHandler.h
Using information in the lecture and the first event handler tutorial, implemented eventHandler.h and an EventHandler class that currently only deals with the WindowCloseEvent.

application.h
In protected created a handler in a bool onClose(WindowCloseEvent& event)

application.cpp
Added the onClose() with bools changing the handleEvent bool in event.h, the m_running with applicaiton.h and returning that the event is handled.
Then tested by having the window close after a second with the Application::run().
Did as expected so at the moment working as desired.