windowEvents.h & mouseEvents.h
mouseEvents.h
Further refined file, added missing vec2 getMousePosition() to MouseMovementEvent class.

windowEvents.h
Added class WindowEvent to deal with the getCategoryFlags() as a lot of repetition and had the other classes inherit from this.
Added a missing ivec2 getWindowPosition() to the WindowMoveEvent class.
Otherwise cleaned up the code here and made it a little more presentable.

Compiled and ran without errors. Seems to be good to go.