GLFWWindowImplement.cpp & application.h/cpp
Updated the intiWindow(), linked in the event handler, set the eventHandler to the void point in the glfwSetWindowUserPointer(), has the lambda function in to set the callback.
Updated application.h/cpp removing the direct eventHandler, instead accessing it through m_window pointer to the window class.

Now the window closes when X clicked.